housekeeping. added some todos and cleaned up some things. also set some components to be deprecated if nothing breaks. Renamed some routes and components as well to be more descriptive.
