+++ src/components/ClientOnly.jsx
@@ -11,3 +11,8 @@ export default function ClientOnly({ children }) {
   if (!mounted) return null;
   return children;
 }
+
+
+// TODO: this component may not be necessary. It checks if user is logged in
+// but there are other ways to do this that seem to work. 
+// Looked cute, may delete later.
